Refactor:totalSales 필드 타입을 BigDecimal로 변경

+++ src/main/java/com/fairing/fairplay/statistics/dto/event/EventStatsOverviewResponseDto.java
@@ -2,6 +2,8 @@ package com.fairing.fairplay.statistics.dto.event;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -10,6 +12,6 @@ import lombok.*;
 public class EventStatsOverviewResponseDto {
     private Long totalUsers;
     private Long totalReservations;
-    private Long totalSales;
+    private BigDecimal totalSales;
     private Long totalEvents;
 }
