+++ src/main/java/com/fairing/fairplay/statistics/service/event/EventComparisonService.java
@@ -10,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
@@ -123,7 +124,7 @@ public class EventComparisonService {
         return EventStatsOverviewResponseDto.builder()
                 .totalUsers(totalUsers)
                 .totalReservations(totalReservations)
-                .totalSales(totalSales)
+                .totalSales(BigDecimal.valueOf(totalSales))
                 .totalEvents((long) allStats.size())
                 .build();
     }
